Only move single tests that are failing
